+++ examples/solidstart/src/middleware/index.ts
@@ -1,0 +1,4 @@
+import { fedifyMiddleware } from "@fedify/solidstart";
+import federation from "../lib/federation";
+
+export default fedifyMiddleware(federation);